Separated littleGuy into a new file
Took littleGuy functions out of the homeScreen.js file so that they can
be imported into other files separately.
