+++ little-guy-creator-app/.gitignore
@@ -8,6 +8,7 @@ node_modules/
 dist/
 web-build/
 expo-env.d.ts
+.idea/
 
 # Native
 *.orig.*
